+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add basic `hooks`
 - chore: Initial scaffolding